+++ www/src/pages-markdown/whatsnew.md
@@ -7,6 +7,7 @@ header: true
 
 ### 0.8.72-3.1.1 (In Progress)
 
+ - Spirit Saber now always shows you the exact card that had the aftermath when it triggered, even if it was transformed later. (1657)
  - Archaeologist's Taletellers now correctly looks for minions played this turn, not minions that were played and killed this turn. (1652)
  - The Monk class has been retired. Its class and cards have been made uncollectible by default. "So long and thanks for all deflect." (1635, 1636)
  - Storyteller's Overflowing Energy now correctly only targets one instance of the lowest-cost card in your hand. (1642)